redesign the task management part
Set a default task for task chosing. Once new task was chosen, the default one is designed to pend(not add the code yet). When running task deleted, the default task will    be running again until task was chosen to run.
All mentioned above not finished yet in this commit.
To be continued.